feat: update the method for retrieve the ebook's page indicator to ensure compatibility with older books
Some older ebooks display the page indicator as "of <total page>" instead of "<current page> of <total page>".
To accommodate this, the method for retrieving the page indicator in the ebook parser has been updated.
As a result, `current page` can no longer be assumed to always exist, so related functionality has been removed from the parser.

Instead, the `ebook_service` now determines the current page based on the loop index.
This approach makes the logic for detecting the last page less robust since it relies on the index rather than directly retrieving values from the ebook viewer.
However, ensuring compatibility with older ebooks is a higher priority, making this trade-off acceptable.